requirements.txt
asgiref==3.7.2
certifi==2022.12.7
dj-database-url==1.2.0
Django==3.2
django-heroku==0.3.1
django-taggit==3.1.0
gunicorn==21.2.0
Markdown==3.1.1
packaging==23.1
psycopg2==2.9.6
python-decouple==3.7
pytz==2020.1
sqlparse==0.3.1
tzdata==2023.3
whitenoise==6.5.0
wincertstore==0.2
